% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss4dS2.R
\name{ss4dS2}
\alias{ss4dS2}
\title{The required sample size for estimating a single difference of variances}
\usage{
ss4dS2(N, S2.1, S2.2, DEFF = 1, conf = 0.95, cve = 0.05, me,
  plot = FALSE)
}
\arguments{
\item{N}{The maximun population size between the groups (strata) that we want to compare.}

\item{S2.1}{The value of the first estimated variance}

\item{S2.2}{The value of the second estimated variance.}

\item{DEFF}{The design effect of the sample design. By default \code{DEFF = 1}, which corresponds to a simple random sampling design.}

\item{conf}{The statistical confidence. By default conf = 0.95. By default \code{conf = 0.95}.}

\item{cve}{The maximun coeficient of variation that can be allowed for the estimation.}

\item{me}{The maximun absolute margin of error that can be allowed for the estimation.}

\item{plot}{Optionally plot the errors (cve and margin of error) against the sample size.}
}
\description{
This function returns the minimum sample size required for estimating a single difference of variances subjecto to predefined errors.
}
\details{
Note that the minimun sample size to achieve a particular margin of error \eqn{\varepsilon} is defined by: 
\deqn{n = \frac{n_0}{\frac{(N-3)(N-2)}{N(2N-7)}+\frac{n_0}{N}}}
Where \deqn{n_0=\frac{z^2_{1-\frac{\alpha}{2}}S^2}{\varepsilon^2}}
and
\deqn{S^2=(S^4_1+S^4_2)DEFF}
Also note that the minimun sample size to achieve a particular coefficient of variation \eqn{cve} is defined by:
\deqn{n = \frac{S^2}{(S^2_1-S^2_2)^2cve^2\frac{(N-3)(N-2)}{N(2N-7)}+\frac{S^2}{N}}}
}
\examples{
ss4dS2(N=100000, S2.1=100, S2.2=150, cve=0.05, me=10)
ss4dS2(N=100000, S2.1=100, S2.2=150, cve=0.05, me=10, plot=TRUE)

#############################
# Example with BigLucy data #
#############################
data(BigLucy)
attach(BigLucy)

N1 <- table(SPAM)[1]
N2 <- table(SPAM)[2]
N <- max(N1,N2)
P1 <- prop.table(table(SPAM))[1]
P2 <- prop.table(table(SPAM))[2]
# The minimum sample size for simple random sampling
ss4dp(N, P1, P2, DEFF=1, conf=0.99, cve=0.03, me=0.03, plot=TRUE)
# The minimum sample size for a complex sampling design
ss4dp(N, P1, P2, DEFF=3.45, conf=0.99, cve=0.03, me=0.03, plot=TRUE)
}
\author{
Hugo Andres Gutierrez Rojas <hugogutierrez at usantotomas.edu.co> and Hanwen Zhang <hanwenzhang at usantotomas.edu.co>
}
\references{
Gutierrez, H. A. (2009), \emph{Estrategias de muestreo: Diseno de encuestas y estimacion de parametros}. Editorial Universidad Santo Tomas
}
\seealso{
\code{\link{e4p}}
}

