% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss4dS2H.R
\name{ss4dS2H}
\alias{ss4dS2H}
\title{The required sample size for testing a null hyphotesis for a single difference of variances}
\usage{
ss4dS2H(N, S2.1, S2.2, D, DEFF = 1, conf = 0.95, power = 0.8,
  plot = FALSE)
}
\arguments{
\item{N}{The maximun population size between the groups (strata) that we want to compare.}

\item{S2.1}{The value of the first estimated variance.}

\item{S2.2}{The value of the second estimated variance}

\item{D}{The minimun effect to test.}

\item{DEFF}{The design effect of the sample design. By default \code{DEFF = 1}, which corresponds to a simple random sampling design.}

\item{conf}{The statistical confidence. By default \code{conf = 0.95}.}

\item{power}{The statistical power. By default \code{power = 0.80}.}

\item{plot}{Optionally plot the effect against the sample size.}
}
\description{
This function returns the minimum sample size required for testing a null hyphotesis regarding a single difference of variances
}
\details{
We assume that it is of interest to test the following set of hyphotesis:
\deqn{H_0: S^2_1 - S^2_2 = 0 \ \ \ \ vs. \ \ \ \ H_a: S^2_1 - S^2_2 = D > 0 }
Note that the minimun sample size, restricted to the predefined power \eqn{\beta} and confidence \eqn{1-\alpha}, is defined by: 
\deqn{n = \frac{S^2}{\frac{D^2}{(z_{1-\alpha} + z_{\beta})^2}\frac{(N-3)(N-2)}{N(2N-7)}+\frac{S^2}{N}}}
Where \eqn{S^2=(S^4_1+S^4_2)DEFF}.
}
\examples{
ss4dS2H(N=100000, S2.1=100, S2.2=120, D=20)
ss4dS2H(N=100000, S2.1=100, S2.2=120, D=20, power = 0.9)
ss4dS2H(N=100000, S2.1=100, S2.2=120, D=20, plo=TRUE)

#############################
# Example with BigLucy data #
#############################
data(BigLucy)
attach(BigLucy)

N1 <- table(SPAM)[1]
N2 <- table(SPAM)[2]
N <- max(N1,N2)
P1 <- prop.table(table(SPAM))[1]
P2 <- prop.table(table(SPAM))[2]

# The minimum sample size for testing 
# H_0: P_1 - P_2 = 0   vs.   H_a: P_1 - P_2 = D = 0.05
D = 0.05  
ss4dpH(N, P1, P2, D, DEFF = 2, plot=TRUE)

# The minimum sample size for testing 
# H_0: P - P_0 = 0   vs.   H_a: P - P_0 = D = 0.02
D = 0.01
ss4dpH(N, P1, P2, D, conf = 0.99, power = 0.9, DEFF = 3.45, plot=TRUE)
}
\author{
Hugo Andres Gutierrez Rojas <hugogutierrez at usantotomas.edu.co> and Hanwen Zhang <hanwenzhang at usantotomas.edu.co>
}
\references{
Gutierrez, H. A. (2009), \emph{Estrategias de muestreo: Diseno de encuestas y estimacion de parametros}. Editorial Universidad Santo Tomas
}
\seealso{
\code{\link{ss4pH}}
}

