% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss4HHSp.R
\name{ss4HHSp}
\alias{ss4HHSp}
\title{Sample Sizes for Household Surveys in Two-Stages for Estimating Single Proportions}
\usage{
ss4HHSp(N, M, r, b, rho, P, delta, conf, m)
}
\arguments{
\item{N}{The population size.}

\item{M}{Number of clusters in the population.}

\item{r}{Percentage of people within the subpopulation of interest.}

\item{b}{Average household size (number of members).}

\item{rho}{The Intraclass Correlation Coefficient.}

\item{P}{The value of the estimated proportion.}

\item{delta}{The maximun margin of error that can be allowed for the estimation.}

\item{conf}{The statistical confidence. By default \code{conf = 0.95}.}

\item{m}{(vector) Number of households selected within PSU.}
}
\value{
This function returns a grid of possible sample sizes. 
The first column represent the design effect,
the second column is the number of clusters to be selected, 
the third column is the number of units to be selected inside the clusters, 
and finally, the last column indicates the full sample size induced by this particular strategy.
}
\description{
This function computes a grid of possible sample sizes for estimating single proportions under two-stage sampling designs.
}
\details{
In two-stage (2S) sampling, the design effect is defined by
\deqn{DEFF = 1 + (\bar{m}-1)\rho} 
Where \eqn{\rho} is defined as the intraclass correlation coefficient,  
\eqn{\bar{m}} is the average sample size of units selected inside each cluster. 
The relationship of the full sample size of the two stage design (2S) with the 
simple random sample (SI) design is given by
\deqn{ n_{2S} =  n_{SI}*DEFF}
}
\examples{
N <- 50000000
M <- 30000
r <- 1
b <- 3.5
m <- 10
rho <- 0.034
P <- 0.04
delta <- 0.05
conf <- 0.9

m <- c(1:10)

ss4HHSp(N = N, M = M, r = r, b = b, 
rho = rho, P = P, delta = delta, conf = conf,
m = m)

}
\references{
Gutierrez, H. A. (2009), \emph{Estrategias de muestreo: Diseno de encuestas y estimacion de parametros}. Editorial Universidad Santo Tomas
}
\seealso{
\code{\link{ICC}}
}
\author{
Hugo Andres Gutierrez Rojas <hugogutierrez at gmail.com>
}
