% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss4S2.R
\name{ss4S2}
\alias{ss4S2}
\title{The required sample size for estimating a single variance}
\usage{
ss4S2(N, DEFF = 1, conf = 0.95, cve = 0.05, rme = 0.03, plot = FALSE)
}
\arguments{
\item{N}{The population size.}

\item{DEFF}{The design effect of the sample design. By default \code{DEFF = 1}, which corresponds to a simple random sampling design.}

\item{conf}{The statistical confidence. By default conf = 0.95. By default \code{conf = 0.95}.}

\item{cve}{The maximun coeficient of variation that can be allowed for the estimation.}

\item{rme}{The maximun relative margin of error that can be allowed for the estimation.}

\item{plot}{Optionally plot the errors (cve and margin of error) against the sample size.}
}
\description{
This function returns the minimum sample size required for estimating variance in a finite population subjecto to predefined errors.
}
\details{
Note that the minimun sample size to achieve a particular relative margin of error \eqn{\varepsilon} is defined by: 
\deqn{n = \frac{n_0}{\frac{(N-3)(N-2)}{N(2N-7)}+\frac{n_0}{N}}}
Where \deqn{n_0=\frac{z^2_{1-\frac{\alpha}{2}}DEFF}{\varepsilon^2}}
Also note that the minimun sample size to achieve a particular coefficient of variation \eqn{cve} is defined by:
\deqn{n = \frac{S^2}{p^2cve^2+\frac{S^2}{N}}}
}
\examples{
ss4S2(N=100000, DEFF=1, conf=0.95, cve=0.05, rme=0.03)
ss4S2(N=100000, DEFF=1, conf=0.95, cve=0.05, rme=0.03, plot=TRUE)
ss4S2(N=100000, DEFF=1, conf=0.95, cve=0.05, rme=0.05, plot=TRUE)

##########################
# Example with Lucy data #
##########################

data(Lucy)
attach(Lucy)
N <- nrow(Lucy)
# The population variance of the variable Income
S2 <- var(Lucy$Income)
# The minimum sample size for simple random sampling
ss4p(N, p, DEFF=1, conf=0.99, cve=0.03, me=0.03, plot=TRUE)
# The minimum sample size for a complex sampling design
ss4p(N, p, DEFF=3.45, conf=0.99, cve=0.03, me=0.03, plot=TRUE)
}
\author{
Hugo Andres Gutierrez Rojas <hugogutierrez at usantotomas.edu.co> and Hanwen Zhang <hanwenzhang at usantotomas.edu.co>
}
\references{
Gutierrez, H. A. (2009), \emph{Estrategias de muestreo: Diseno de encuestas y estimacion de parametros}. Editorial Universidad Santo Tomas
}
\seealso{
\code{\link{ss4S2H}}
}

